requirements.txt
cocos2d==0.6.4
coverage==4.2
psychic-engine==0.0.0
py==1.4.31
pyglet==1.2.4
pytest==2.9.2
pytest-capturelog==0.7
pytest-mock==1.1
requests==2.10.0
six==1.10.0
wit==4.0.0
